Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. Although the name of the repository does match the sketch name, when GitHub's popular Clone or download > Download ZIP feature is used to download a repository the branch/release/commit name is appended to the folder name, causing a mismatch.

When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "Arduino-MoonLite-focuser.ino" needs to be inside a sketch folder named "Arduino-MoonLite-focuser".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file Arduino-MoonLite-focuser.ino to the new folder, leaving behind Config.h. This causes compilation of the sketch to fail:

Arduino-MoonLite-focuser-master\Arduino-MoonLite-focuser\Arduino-MoonLite-focuser.ino:20:20: fatal error: Config.h: No such file or directory

 #include "Config.h"